refactor: move DateFormatter.postFormat(date:) to its own file
